app/requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
Django==3.0.6
django-appconf==1.0.4
django-crispy-forms==1.9.1
django-cryptography==1.0
mysqlclient==1.4.6
Pillow==7.1.2
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
